call check_status from one way and move Pipeline class
