+++ tasks/maintenance/check.py
@@ -4,8 +4,9 @@ import traceback
 
 import pywikibot
 
+from core.utils.helpers import check_status
 from core.utils.sqlite import create_database_table, maintenance_db_name, get_articles
-from module import process_article, check_status
+from module import process_article
 
 
 def read(thread_number):
@@ -14,7 +15,7 @@ def read(thread_number):
         site = pywikibot.Site()
         conn, cursor = create_database_table(maintenance_db_name)
         rows = get_articles(cursor, thread_number)
-        if len(rows) > 0 and check_status():
+        if len(rows) > 0 and check_status("مستخدم:LokasBot/إيقاف مهمة صيانة المقالات"):
             for row in rows:
                 print(row)
                 process_article(site, cursor, conn, id=row[0], title=row[1], thread_number=thread_number)
